+++ server/src/main/java/org/realityforge/replicant/server/transport/SystemMetaData.java
@@ -1,7 +1,6 @@
 package org.realityforge.replicant.server.transport;
 
 import java.util.Arrays;
-import java.util.List;
 import java.util.Objects;
 import javax.annotation.Nonnull;
 import org.realityforge.replicant.server.ChannelAddress;
@@ -11,7 +10,7 @@ public final class SystemMetaData
   @Nonnull
   private final String _name;
   @Nonnull
-  private final List<ChannelMetaData> _channels;
+  private final ChannelMetaData[] _channels;
 
   public SystemMetaData( @Nonnull final String name, @Nonnull final ChannelMetaData... channels )
   {
@@ -26,7 +25,7 @@ public final class SystemMetaData
       }
     }
     _name = Objects.requireNonNull( name );
-    _channels = Arrays.asList( Objects.requireNonNull( channels ) );
+    _channels = channels;
   }
 
   @Nonnull
@@ -35,10 +34,9 @@ public final class SystemMetaData
     return _name;
   }
 
-  @Nonnull
-  public List<ChannelMetaData> getChannels()
+  public int getChannelCount()
   {
-    return _channels;
+    return _channels.length;
   }
 
   @Nonnull
@@ -53,10 +51,10 @@ public final class SystemMetaData
   @Nonnull
   public ChannelMetaData getChannelMetaData( final int channelId )
   {
-    if ( channelId >= _channels.size() || channelId < 0 )
+    if ( channelId >= _channels.length || channelId < 0 )
     {
       throw new NoSuchChannelException( channelId );
     }
-    return _channels.get( channelId );
+    return _channels[ channelId ];
   }
 }